fix: unit tests for ProjectTemplates and ProductTemplates
This issue become visible after fixing the expectObj() method.
The thing here is that when we use .empty(null) in Joi then if property is passed as "null" it's stripped from the object.
As a result, we pass an object with "null" value, but we are trying to save in ES the object without value.
This leads to the unit test fails, as saved value doesn't match the object which has been send in the message.