feat: disable indexing objects on milestone and phase create and update
Temporary disable indexing these objects as for now we keep the logic for this inside Project Service.
It's because creating or updating a phase or milestone may cause cascading updates of other phases or milestones.
In such cases in Project Service we are doing one ES index call instead of multiple calls.
Otherwise ES may fail with error `version conflict`.
This would be turned on back, as soon as we get rid of such cascading updates inside Project Service.